Move testConstants.ts file to test/
